# openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: RESTful API for managing tasks and user assignments
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current user
      description: Get authenticated user information
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations:
    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization with auto-generated invite code
      operationId: createOrganization
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: My Team
      responses:
        "201":
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags:
        - Organizations
      summary: List organizations
      description: Get all organizations the user is a member of
      operationId: listOrganizations
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of organizations with user role
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrganizationWithRole"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations/join:
    post:
      tags:
        - Organizations
      summary: Join organization
      description: Join an organization using invite code
      operationId: joinOrganization
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - invite_code
              properties:
                invite_code:
                  type: string
                  example: "5dc6-6411-e229"
      responses:
        "200":
          description: Successfully joined organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully joined organization
                  organization:
                    $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Invalid invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Already a member of this organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization
      description: Get organization details including members
      operationId: getOrganization
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found or access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Organizations
      summary: Update organization
      description: Update organization name (owner only)
      operationId: updateOrganization
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Updated Team Name
      responses:
        "200":
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only organization owners can perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Organizations
      summary: Delete organization
      description: Delete organization and all its tasks (owner only)
      operationId: deleteOrganization
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Organization deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Organization deleted successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only organization owners can perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations/{id}/regenerate-code:
    post:
      tags:
        - Organizations
      summary: Regenerate invite code
      description: Generate new invite code for organization (owner only)
      operationId: regenerateInviteCode
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Invite code regenerated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only organization owners can perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organizations/{id}/members/{user_id}:
    delete:
      tags:
        - Organizations
      summary: Remove member
      description: Remove a member from organization (owner only)
      operationId: removeMember
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: integer
            format: int64
        - name: user_id
          in: path
          required: true
          description: User ID to remove
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Member removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Member removed successfully
        "400":
          description: Cannot remove yourself
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only organization owners can perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        created_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
    Organization:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: My Team
        invite_code:
          type: string
          example: "5dc6-6411-e229"
    OrganizationWithRole:
      allOf:
        - $ref: "#/components/schemas/Organization"
        - type: object
          required:
            - role
          properties:
            role:
              type: string
              enum: [owner, member]
              example: owner
    OrganizationMember:
      type: object
      required:
        - user
        - role
        - joined_at
      properties:
        user:
          $ref: "#/components/schemas/User"
        role:
          type: string
          enum: [owner, member]
          example: member
        joined_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
    OrganizationDetail:
      allOf:
        - $ref: "#/components/schemas/Organization"
        - type: object
          required:
            - members
            - your_role
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/OrganizationMember"
            your_role:
              type: string
              enum: [owner, member]
              example: owner
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Error message describing what went wrong